logging eachievement unlocked :)
Instead of using many print statements for debugging, use logger.debug:
unlike the print statements, which you will have to delete or comment
out later, the logger.debug statements can remain intact in the source
code and remain dormant until you need them again. At that time, the
only change that needs to happen is to modify the severity level of the
logger and/or handler to debug.
https://docs.python.org/2/howto/logging-cookbook.html#multiple-handlers-and-formatters